AuthService/requirements.txt
fastapi
uvicorn
firebase-admin
pydantic
httpx
pytest
python-dotenv